// frontend/generated/models/TransactionTypeDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Freelance Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const TransactionTypeDto = {
    Deposit: 'DEPOSIT',
    Withdrawal: 'WITHDRAWAL',
    EscrowDeposit: 'ESCROW_DEPOSIT',
    EscrowRelease: 'ESCROW_RELEASE',
    EscrowRefund: 'ESCROW_REFUND'
} as const;
export type TransactionTypeDto = typeof TransactionTypeDto[keyof typeof TransactionTypeDto];


export function instanceOfTransactionTypeDto(value: any): boolean {
    for (const key in TransactionTypeDto) {
        if (Object.prototype.hasOwnProperty.call(TransactionTypeDto, key)) {
            if (TransactionTypeDto[key as keyof typeof TransactionTypeDto] === value) {
                return true;
            }
        }
    }
    return false;
}

export function TransactionTypeDtoFromJSON(json: any): TransactionTypeDto {
    return TransactionTypeDtoFromJSONTyped(json, false);
}

export function TransactionTypeDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): TransactionTypeDto {
    return json as TransactionTypeDto;
}

export function TransactionTypeDtoToJSON(value?: TransactionTypeDto | null): any {
    return value as any;
}

export function TransactionTypeDtoToJSONTyped(value: any, ignoreDiscriminator: boolean): TransactionTypeDto {
    return value as TransactionTypeDto;
}